moved template validation to src/utils/templateValidation.ts
