+++ src/util/templateValidation.ts
